Improvements to release CI (#6)
* Implemented a new workflow for releasing to pypi

* Fixed a portability issue with printf statements.
The size of npy_intp equals the size of the pointer on a given architecture. On 32-bit systems that equals sizeof(int), but on 64-bit systems it equals sizeof(long int). In consequence, the printf formatter needs to be either %d or %ld; to circumvent that, npy_intp is now cast to int before printing.